requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.16.0
click==8.1.2
colorama==0.4.6
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
pycparser==2.21
pyftpdlib==1.5.9
PyPDF2==3.0.1
python-dotenv==1.0.1
pdf-manipulation>=0.0.2
setuptools==69.5.1
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
Werkzeug==2.1.1
wheel==0.43.0
WTForms==3.0.1
zipp==3.8.0
